requirements.txt
Django==2.2
djangorestframework==3.13.0
gunicorn==21.2.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
